Create a new component that holds the board and import it for left/right boards
