+++ src/components/Board/Board.jsx
@@ -1,15 +1,15 @@
 import React, { useState } from 'react';
+import MainBoard from '../MainBoard/MainBoard';
 
 const Board = () => {
   const [squares, setSquares] = useState(Array(9).fill(null));
   const [xIsNext, setXIsNext] = useState(true);
 
-  // Handle a square click
   const handleClick = (i) => {
-    // Create a copy of the squares array
+    // Create a copy of the squares array - to not manipulate or cause an error in the main one
     const newSquares = squares.slice();
 
-    // If the square is already filled or there's a winner, return early
+    // If the square is already filled, return early
     if (newSquares[i]) {
       return;
     }
@@ -20,31 +20,15 @@ const Board = () => {
     // Update squares state
     setSquares(newSquares);
 
-    // Log the player who will play next
-    console.log(`The next player will be using: ${xIsNext ? 'O' : 'X'}`);
-
     // Switch to the next player
     setXIsNext(!xIsNext);
-  };
-
-  const renderSquare = (i) => {
-    return (
-      <button
-        className="w-16 h-16 border border-gray-400 flex items-center justify-center text-2xl"
-        onClick={() => handleClick(i)}
-      >
-        {squares[i]}
-      </button>
-    );
   };
 
   const status = `Next player: ${xIsNext ? 'Player 1 (X)' : 'Player 2 (O)'}`;
 
   return (
     <div className="flex flex-col items-center justify-center h-screen mt-8">
-      {/* Player 1 and Player 2 graphics */}
       <div className="flex justify-between w-full max-w-[800px] mb-8">
-        {/* Player 1 graphic */}
         <div className={`text-2xl ${xIsNext ? 'text-blue-600 font-bold' : 'text-gray-400'}`}>
           <div className="flex flex-col items-center mt-[85px]">
             <p>Player 1</p>
@@ -54,7 +38,6 @@ const Board = () => {
           </div>
         </div>
 
-        {/* Player 2 graphic */}
         <div className={`text-2xl ${!xIsNext ? 'text-red-600 font-bold' : 'text-gray-400'}`}>
           <div className="flex flex-col items-center mt-[85px]">
             <p>Player 2</p>
@@ -65,51 +48,15 @@ const Board = () => {
         </div>
       </div>
 
-      {/* Status message */}
       <div className="mb-4 text-2xl font-bold">{status}</div>
 
-      {/* Game board */}
       <div className="flex justify-center w-full max-w-[800px]">
-        {/* Left Board */}
-        <div className="flex flex-col items-center justify-center w-1/2 p-4">
-          <div className="flex">
-            {renderSquare(0)}
-            {renderSquare(1)}
-            {renderSquare(2)}
-          </div>
-          <div className="flex">
-            {renderSquare(3)}
-            {renderSquare(4)}
-            {renderSquare(5)}
-          </div>
-          <div className="flex">
-            {renderSquare(6)}
-            {renderSquare(7)}
-            {renderSquare(8)}
-          </div>
-        </div>
+        <MainBoard squares={squares} handleClick={handleClick} />
 
-        {/* Middle Line */}
+        {/* Separate boards with a line */}
         <div className="w-1 bg-black h-full"></div>
 
-        {/* Right Board (Same as Left Board) */}
-        <div className="flex flex-col items-center justify-center w-1/2 p-4">
-          <div className="flex">
-            {renderSquare(0)}
-            {renderSquare(1)}
-            {renderSquare(2)}
-          </div>
-          <div className="flex">
-            {renderSquare(3)}
-            {renderSquare(4)}
-            {renderSquare(5)}
-          </div>
-          <div className="flex">
-            {renderSquare(6)}
-            {renderSquare(7)}
-            {renderSquare(8)}
-          </div>
-        </div>
+        <MainBoard squares={squares} handleClick={handleClick} />
       </div>
     </div>
   );
